traj utils
