+++ examples/datasets/traj.py
@@ -31,6 +31,54 @@ def focus_point_fn(poses: np.ndarray) -> np.ndarray:
     return focus_pt
 
 
+def average_pose(poses: np.ndarray) -> np.ndarray:
+    """New pose using average position, z-axis, and up vector of input poses."""
+    position = poses[:, :3, 3].mean(0)
+    z_axis = poses[:, :3, 2].mean(0)
+    up = poses[:, :3, 1].mean(0)
+    cam2world = viewmatrix(z_axis, up, position)
+    return cam2world
+
+
+def generate_spiral_path(
+    poses,
+    bounds,
+    n_frames=120,
+    n_rots=2,
+    zrate=0.5,
+    spiral_scale_f=1.0,
+    spiral_scale_r=1.0,
+    focus_distance=0.75,
+):
+    """Calculates a forward facing spiral path for rendering."""
+    # Find a reasonable 'focus depth' for this dataset as a weighted average
+    # of conservative near and far bounds in disparity space.
+    near_bound = bounds.min()
+    far_bound = bounds.max()
+    # All cameras will point towards the world space point (0, 0, -focal).
+    focal = 1 / (((1 - focus_distance) / near_bound + focus_distance / far_bound))
+    focal = focal * spiral_scale_f
+
+    # Get radii for spiral path using 90th percentile of camera positions.
+    positions = poses[:, :3, 3]
+    radii = np.percentile(np.abs(positions), 90, 0)
+    radii = radii * spiral_scale_r
+    radii = np.concatenate([radii, [1.0]])
+
+    # Generate poses for spiral path.
+    render_poses = []
+    cam2world = average_pose(poses)
+    up = poses[:, :3, 1].mean(0)
+    for theta in np.linspace(0.0, 2.0 * np.pi * n_rots, n_frames, endpoint=False):
+        t = radii * [np.cos(theta), -np.sin(theta), -np.sin(theta * zrate), 1.0]
+        position = cam2world @ t
+        lookat = cam2world @ [0, 0, -focal, 1.0]
+        z_axis = position - lookat
+        render_poses.append(viewmatrix(z_axis, up, position))
+    render_poses = np.stack(render_poses, axis=0)
+    return render_poses
+
+
 def generate_ellipse_path_z(
     poses: np.ndarray,
     n_frames: int = 120,
