--- examples/datasets/traj_utils.py
@@ -1,67 +1,0 @@
-import numpy as np
-
-# Constants for generate_spiral_path():
-NEAR_STRETCH = 0.9  # Push forward near bound for forward facing render path.
-FAR_STRETCH = 5.0  # Push back far bound for forward facing render path.
-FOCUS_DISTANCE = 0.75  # Relative weighting of near, far bounds for render path.
-
-
-def generate_spiral_path(
-    poses,
-    bounds,
-    n_frames=120,
-    n_rots=2,
-    zrate=0.5,
-    spiral_scale_f=1.0,
-    spiral_scale_r=1.0,
-):
-    """Calculates a forward facing spiral path for rendering."""
-    # Find a reasonable 'focus depth' for this dataset as a weighted average
-    # of conservative near and far bounds in disparity space.
-    near_bound = bounds.min() * NEAR_STRETCH
-    far_bound = bounds.max() * FAR_STRETCH
-    # All cameras will point towards the world space point (0, 0, -focal).
-    focal = 1 / (((1 - FOCUS_DISTANCE) / near_bound + FOCUS_DISTANCE / far_bound))
-    focal = focal * spiral_scale_f
-
-    # Get radii for spiral path using 90th percentile of camera positions.
-    positions = poses[:, :3, 3]
-    radii = np.percentile(np.abs(positions), 90, 0)
-    radii = radii * spiral_scale_r
-    radii = np.concatenate([radii, [1.0]])
-
-    # Generate poses for spiral path.
-    render_poses = []
-    cam2world = average_pose(poses)
-    up = poses[:, :3, 1].mean(0)
-    for theta in np.linspace(0.0, 2.0 * np.pi * n_rots, n_frames, endpoint=False):
-        t = radii * [np.cos(theta), -np.sin(theta), -np.sin(theta * zrate), 1.0]
-        position = cam2world @ t
-        lookat = cam2world @ [0, 0, -focal, 1.0]
-        z_axis = position - lookat
-        render_poses.append(viewmatrix(z_axis, up, position))
-    render_poses = np.stack(render_poses, axis=0)
-    return render_poses
-
-
-def average_pose(poses):
-    """New pose using average position, z-axis, and up vector of input poses."""
-    position = poses[:, :3, 3].mean(0)
-    z_axis = poses[:, :3, 2].mean(0)
-    up = poses[:, :3, 1].mean(0)
-    cam2world = viewmatrix(z_axis, up, position)
-    return cam2world
-
-
-def viewmatrix(lookdir, up, position):
-    """Construct lookat view matrix."""
-    vec2 = normalize(lookdir)
-    vec0 = normalize(np.cross(up, vec2))
-    vec1 = normalize(np.cross(vec2, vec0))
-    m = np.stack([vec0, vec1, vec2, position], axis=1)
-    return m
-
-
-def normalize(x):
-    """Normalization helper function."""
-    return x / np.linalg.norm(x)
